+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Released on Wednesday, September 9 2026.
 
 - Improved support for clipping and `overflow` declarations
+- Fixed painting of scrolled content
 - Added support for rounded corners
 - Added support for `text-shadow` declarations
 - Added support for `box-shadow` declarations